Fluid SDF working yooooooo

+++ FluidSolverCore/src/steps/Advection.cpp
@@ -82,7 +82,7 @@ void Advection::execute(std::vector<MarkerParticle>& markerParticles, const Stag
 		Vec2f averageVel = (currentVel + finalVel) / 2;
 		Vec2f newPos = particle.position + averageVel / dx * timeStep;
 
-		const CellData& finalCell = cellData.getValue(std::floor(newPos.x), std::floor(newPos.y));
+		const CellData& finalCell = cellData.getValue(static_cast<int>(std::floor(newPos.x)), static_cast<int>(std::floor(newPos.y)));
 		if (finalCell.cellType != CellType::Solid)
 			particle.position = newPos;
 	}
